+++ driver-sync/src/test/functional/com/mongodb/client/ClientMetadataProseTest.java
@@ -19,7 +19,7 @@ package com.mongodb.client;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoDriverInformation;
 import com.mongodb.client.internal.InternalMongoClients;
-import com.mongodb.internal.connection.InternalMongoClientSettings;
+import com.mongodb.internal.InternalMongoClientSettings;
 import com.mongodb.lang.Nullable;
 
 public class ClientMetadataProseTest extends AbstractClientMetadataProseTest {
